requirements.txt
python-telegram-bot==20.6
google-api-python-client==2.108.0
google-auth==2.23.4
google-generativeai==0.3.1
pyyaml==6.0.1
